dom/base/test/meta_viewport/test_meta_viewport_auto_size_by_fixed_width_and_device_height.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Fixed width and device-height disables autoSize</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=400,height=device-height">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=400, height=device-height</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function fixed_width_and_device_height() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      ok(!info.autoSize,
         "Fixed width should disable autoSize even if height is device-height");
    });
  </script>
</body>
</html>
